Move ocl._declare_function into core.utils.cgutils.extra
